Inline AbstractRowBlock and AbstractSingleRowBlock
