+++ core/trino-spi/src/test/java/io/trino/spi/block/TestRowBlock.java
@@ -26,7 +26,7 @@ public class TestRowBlock
     public void testFieldBlockOffsetsIsNullWhenThereIsNoNullRow()
     {
         Block fieldBlock = new ByteArrayBlock(1, Optional.empty(), new byte[]{10});
-        AbstractRowBlock rowBlock = (RowBlock) RowBlock.fromFieldBlocks(1, Optional.empty(), new Block[] {fieldBlock});
+        RowBlock rowBlock = (RowBlock) RowBlock.fromFieldBlocks(1, Optional.empty(), new Block[] {fieldBlock});
         // Blocks should discard the offset mask during creation if no values are null
         assertThat(rowBlock.getFieldBlockOffsets()).isNull();
     }
@@ -35,7 +35,7 @@ public class TestRowBlock
     public void testFieldBlockOffsetsIsNotNullWhenThereIsNullRow()
     {
         Block fieldBlock = new ByteArrayBlock(1, Optional.empty(), new byte[]{10});
-        AbstractRowBlock rowBlock = (RowBlock) RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {true}), new Block[] {fieldBlock});
+        RowBlock rowBlock = (RowBlock) RowBlock.fromFieldBlocks(1, Optional.of(new boolean[] {true}), new Block[] {fieldBlock});
         // Blocks should not discard the offset mask during creation if no values are null
         assertThat(rowBlock.getFieldBlockOffsets()).isNotNull();
     }
